@@ -1,10 +1,13 @@
-// swift-tools-version:5.5
+// swift-tools-version:6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "SwiftObserver",
+//    platforms: [
+//        .iOS(.v15),
+//    ],
     products: [
         .library(
             name: "SwiftObserver",
