@@ -2,16 +2,16 @@
 
 /// Шаблон Наблюдатель события
 /// Источник связи "один ко многим".
-public protocol EventProtocol {
-    associatedtype Parameter
+public protocol EventProtocol: Sendable {
+    associatedtype Parameter: Sendable
 
     /// Долавление нового слушателя
-    static func += (event: Self, handler: EventObserver<Parameter>)
+    func addObserver(_ observer: EventObserver<Parameter>) async
 }
 
 // MARK: -
 /// База для обработчиков сообщений.
-open class EventObserver<Parameter> {
+open class EventObserver<Parameter: Sendable>: @unchecked Sendable {
     /// Обработать полученное событие.
     /// Возвращает статус true - слушатель готов получать дальнейшие события. false - больше не посылать.
     open func handle(_ value: Parameter) -> Bool {
@@ -24,11 +24,12 @@ open class EventObserver<Parameter> {
 
 // MARK: -
 /// Реализация
-public final class Event<Parameter>: EventProtocol {
+public final class Event<Parameter: Sendable>: EventProtocol, @unchecked Sendable {
     public typealias Observer = EventObserver<Parameter>
 
     /// Cписок обработчиков.
-    private final class Node {
+    /// Потокобезопасность обеспечиваем через actor NodeList
+    private final class Node: @unchecked Sendable {
         var observer: Observer
         var next: Node?
 
@@ -37,11 +38,38 @@ public final class Event<Parameter>: EventProtocol {
             self.next = next
         }
     }
-    private var observers: Node?
-    private var connectionNotifier: (() -> Void)?
-    
-    /// connectedNotifier - опциональный слушатель подключения первого наблюдателя
-    internal init(connectionNotifier: (() -> Void)?) {
+
+    private actor NodeList {
+        var head: Node?
+
+        func add(observer: Observer) async {
+            head = Node(observer: observer, next: head)
+        }
+
+        func notify(_ value: Parameter) async {
+            // Рекурсивный проход по слушателям с удалением отвалившихся.
+            func recursiveWalk(_ node: Node?) -> Node? {
+                guard node != nil else { return nil }
+                var node = node
+                // Схлопываем пустые узлы
+                while let current = node, !current.observer.handle(value) {
+                    node = current.next
+                }
+                if let current = node {
+                    current.next = recursiveWalk(current.next)
+                }
+                return node
+            }
+
+            head = recursiveWalk(head)
+        }
+    }
+
+    private let observers = NodeList()
+    private let connectionNotifier: ((Bool) -> Void)?
+
+    /// connectedNotifier - опциональный слушатель подключения первого наблюдателя и отключения последнего
+    internal init(connectionNotifier: ((Bool) -> Void)?) {
         self.connectionNotifier = connectionNotifier
     }
 
@@ -49,32 +77,19 @@ public final class Event<Parameter>: EventProtocol {
     /// При этом все отвалившиеся слушатели удаляются из списка.
     /// Недоступна для внешнего вызова.
     /// Для внешнего вызова использовать EventSource.
-    /// *returns* true если есть подключения слушателей
-    internal func notifyObservers(_ value: Parameter) -> Bool {
-        // Рекурсивный проход по слушателям с удалением отвалившихся.
-        func recursiveWalk(_ node: Node?) -> Node? {
-            guard node != nil else { return nil }
-            var node = node
-            // Схлопываем пустые узлы
-            while let current = node, !current.observer.handle(value) {
-                node = current.next
-            }
-            if let current = node {
-                current.next = recursiveWalk(current.next)
-            }
-            return node
+    internal func notifyObservers(_ value: Parameter) async {
+        await observers.notify(value)
+        if await self.observers.head == nil {
+            self.connectionNotifier?(false)
         }
-        
-        observers = recursiveWalk(observers)
-        return observers != nil
     }
     
     /// Добавление слушателя. Слушатель добавляется по слабой ссылке. Чтобы убрать слушателя, надо удалить его объект.
     /// Допустимо применять посредника (Observer.Link) для отключения слушателя без удаления целевого боъекта.
-    public static func += (event: Event, observer: Observer) {
-        if event.observers == nil {
-            event.connectionNotifier?()
+    public func addObserver(_ observer: EventObserver<Parameter>) async {
+        if await observers.head == nil {
+            connectionNotifier?(true)
         }
-        event.observers = Node(observer: observer, next: event.observers)
+        await observers.add(observer: observer)
     }
 }
