@@ -1,6 +1,6 @@
 //  Copyright (C) Oleg Bakharev 2021. All Rights Reserved
 
-import XCTest
+import Testing
 import SwiftObserver
 
 private protocol Subject {
@@ -11,57 +11,57 @@ private final class Emitter: Subject {
     private var voidSender = EventSender<Void>()
     var eventVoid: Event<Void> { voidSender.event }
 
-    func send() {
-        voidSender.send()
+    func send() async {
+        await voidSender.send()
     }
 }
 
-private final class Receiver {
+private final class Receiver: Sendable {
     func onVoid() {
         print("Event received")
     }
 }
 
-final class ObserverSandbox: XCTestCase {
+@Suite struct ObserverSandbox {
     // UIControl-like connection. Connection breaks on target release.
-    public func testTargetActionObserver() {
+    @Test func testTargetActionObserver() async throws {
         let emitter = Emitter()
         let receiver = Receiver()
         let subject: Subject = emitter
-        subject.eventVoid += Observer(target: receiver, action: Receiver.onVoid)
-        emitter.send() // "Event received"
+        await subject.eventVoid.addObserver(Observer(target: receiver, action: Receiver.onVoid))
+        await emitter.send() // "Event received"
     }
 
     // UIControl-like connection. Connection breaks on link release.
-    public func testTargetActionLinkObserver() {
+    @Test func testTargetActionLinkObserver() async throws {
         let emitter = Emitter()
         let receiver = Receiver()
         let subject: Subject = emitter
         var mayBeLink: Any?
         do {
             let link = Observer.Link(target: receiver, action: Receiver.onVoid)
-            subject.eventVoid += link
+            await subject.eventVoid.addObserver(link)
             mayBeLink = link
         }
-        XCTAssertNotNil(mayBeLink)
-        emitter.send() // Event received
+        #expect(mayBeLink != nil)
+        await emitter.send() // Event received
         mayBeLink = nil
-        emitter.send() // No output
+        await emitter.send() // No output
     }
 
     // Prmanent closure observer
-    func testPermanentClosure() {
+    @Test func testPermanentClosure() async throws {
         let emitter = Emitter()
         let subject: Subject = emitter
-//        subject.eventVoid += ObserverClosure<Void>() { ... }
-        subject.eventVoid += { // OMG!!!
+//      await subject.eventVoid.addObserver(ObserverClosure<Void>() { ... })
+        await subject.eventVoid.addObserver { // OMG!!!
             print("Event received")
         }
-        emitter.send() // Event received
+        await emitter.send() // Event received
     }
 
     // Disposable closure link. Connection breaks on link release.
-    func testClosureLinkObserevr() {
+    @Test func testClosureLinkObserevr() async throws {
         let emitter = Emitter()
         let subject: Subject = emitter
         var maybeLink: Any?
@@ -69,12 +69,12 @@ final class ObserverSandbox: XCTestCase {
             let link = ObserverClosure.Link {
                 print("Event received")
             }
-            subject.eventVoid += link
+            await subject.eventVoid.addObserver(link)
             maybeLink = link
         }
-        XCTAssertNotNil(maybeLink)
-        emitter.send() // Event received
+        #expect(maybeLink != nil)
+        await emitter.send() // Event received
         maybeLink = nil
-        emitter.send() // No output
+        await emitter.send() // No output
     }
 }
