@@ -1,6 +1,6 @@
-//  Copyright (C) Oleg Bakharev 2021. All Rights Reserved
+//  Copyright (C) Oleg Bakharev 2021-2025. All Rights Reserved
 
-import XCTest
+import Testing
 import SwiftObserver
 
 protocol AsyncSubject {
@@ -13,11 +13,11 @@ actor EmitterActor {
     private var intSender = EventSender<Int>()
 
     func sendVoid() async {
-        voidSender.send()
+        await voidSender.send()
     }
 
     func sendInt() async {
-        intSender.send(0)
+        await intSender.send(0)
     }
 }
 
@@ -26,23 +26,33 @@ extension EmitterActor: AsyncSubject {
     var eventInt: Event<Int> { intSender.event }
 }
 
-
-final class ObserverAsyncTests: XCTestCase {
-
-    func testActorEvents() async {
-        let voidExp = expectation(description: "Void")
-        let intExp = expectation(description: "Int")
-        let e = EmitterActor()
-        let s: AsyncSubject = e
-        await s.eventVoid += {
-            voidExp.fulfill()
-        }
-        await s.eventInt += { _ in
-            intExp.fulfill()
-        }
-        await e.sendVoid()
-        await e.sendInt()
-        await waitForExpectations(timeout: 1)
+actor ReceiverA {
+    func subsribe(_ event: Event<Void>) async {
+        await event.addObserver { }
     }
+}
 
+actor ReceiverB {
+    func subsribe(_ event: Event<Void>) async {
+        await event.addObserver { }
+    }
+}
+
+@Test func testActorEvents() async throws {
+    let e = EmitterActor()
+    let s: AsyncSubject = e
+    var voidReceived = false
+    var intRevceivd = false
+
+    await s.eventVoid.addObserver {
+        voidReceived = true
+    }
+    await s.eventInt.addObserver { _ in
+        intRevceivd = true
+    }
+    await e.sendVoid()
+    await e.sendInt()
+
+    #expect(voidReceived)
+    #expect(intRevceivd)
 }
